(refactor) Move `NoThis` to `util/stamps`
