+++ src/logic/items/Item.js
@@ -1,5 +1,5 @@
 import stampit from "stampit";
-import NoThis from "util/NoThis";
+import NoThis from "util/stamps/NoThis";
 import UniqueId from "util/stamps/UniqueId";
 
 const Item = stampit.compose(NoThis, UniqueId(), stampit({
