+++ src/util/stamps/NoThis.js
